Flags

+++ main.go
@@ -10,21 +10,57 @@ import (
 )
 
 func main() {
+	var dryRun bool
+	var quiet bool
+	var maxAge int
+
 	app := &cli.App{
 		Name:  "git-repo",
 		Usage: "Open the current repo in browser",
-		Action: func(c *cli.Context) (err error) {
-			maxAge := 1
-			branches, err := GetBranchDates()
+		Flags: []cli.Flag{
+			&cli.BoolFlag{
+				Name:        "dry-run",
+				Aliases:     []string{"d"},
+				Usage:       "Dry run to find which branches would be deleted.",
+				Value:       false,
+				Destination: &dryRun,
+			},
+			&cli.BoolFlag{
+				Name:        "quiet",
+				Aliases:     []string{"q"},
+				Usage:       "Hide output.",
+				Value:       false,
+				Destination: &quiet,
+			},
+			&cli.IntFlag{
+				Name:        "age",
+				Aliases:     []string{"a"},
+				Usage:       "Max age: Any branches n+ days old will be deleted. Default: 30",
+				Value:       30,
+				Destination: &maxAge,
+			},
+		},
+		Action: func(c *cli.Context) error {
+			if branches, err := GetBranchDates(); err == nil {
+				for _, staleBranch := range filterForActiveBranches(branches, maxAge) {
+					if !quiet {
+						fmt.Printf("Deleted branch: %s", staleBranch)
+					}
 
-			for _, staleBranch := range filterForActiveBranches(branches, maxAge) {
-				fmt.Println(staleBranch)
-				err = DeleteBranch(staleBranch)
-				if err != nil {
-					return err
+					if dryRun {
+						fmt.Printf("Branch %s is stale", staleBranch)
+						continue
+					}
+
+					err = DeleteBranch(staleBranch)
+					if err != nil {
+						fmt.Fprintf(os.Stderr, "An error occured while deleting branch: %s\n%s", staleBranch, err.Error())
+					}
 				}
+			} else {
+				return err
 			}
-			return
+			return nil
 		},
 	}
 
@@ -38,7 +74,7 @@ func filterForActiveBranches(branches map[string]string, maxAge int) []string {
 	staleBranches := make([]string, 0)
 
 	for branch, date := range branches {
-		lastUpdated, _ := time.Parse("2006-10-25", date)
+		lastUpdated, _ := time.Parse("2006-01-02", date)
 		if time.Now().Sub(lastUpdated).Hours() > float64(maxAge) {
 			staleBranches = append(staleBranches, branch)
 		}
